Speak ingest protocol v2 so a record can carry a Modbus message
v1's record had no spare bit and a 64-byte cap. Each record now names its
kind, `len` is a u16, and the gateway stores a Modbus message with its unit,
function code and CRC verdict beside the CAN rows. The three read endpoints
can name a protocol at last; the gateway accepts v1 for one release so the
daemons can follow it.

+++ crates/wiretap-backend/src/types.rs
@@ -195,7 +195,7 @@ pub struct InventoryEntry {
     pub count: i64,
     pub first_us: i64,
     pub last_us: i64,
-    pub max_dlc: u8,
+    pub max_dlc: u16,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -209,7 +209,7 @@ pub struct FrameBatchRow {
     pub ts_us: i64,
     pub id: u32,
     pub extended: bool,
-    pub dlc: u8,
+    pub dlc: u16,
     pub is_fd: bool,
     pub bus: u8,
     pub dir: String,
